Use loading skeleton on featured site card
